remove code used in dates calculations and superfluous dayPicker
