+++ content/webapp/components/ItemRequestModal/RequestDialog.tsx
@@ -1,11 +1,8 @@
 import { FunctionComponent, FormEvent, useState } from 'react';
-import { useAvailableDates } from './useAvailableDates';
-import { isRequestableDate } from '../../utils/dates';
 import { allowedRequests } from '@weco/common/values/requests';
 import { font } from '@weco/common/utils/classnames';
 import Space from '@weco/common/views/components/styled/Space';
 import RequestingDayPicker from '../RequestingDayPicker/RequestingDayPicker';
-import NewRequestingDayPicker from '../RequestingDayPicker/NewRequestingDayPicker';
 import Button, { ButtonTypes } from '@weco/common/views/components/Buttons';
 import {
   PhysicalItem,
@@ -15,7 +12,6 @@ import styled from 'styled-components';
 import { CTAs, CurrentRequests, Header } from './common';
 import { themeValues } from '@weco/common/views/themes/config';
 import { dateAsValue, dateFromValue } from './format-date';
-import { useToggles } from '@weco/common/server-data/Context';
 
 const PickUpDate = styled(Space).attrs({
   $v: { size: 'l', properties: ['padding-top', 'padding-bottom'] },
@@ -86,11 +82,11 @@ const RequestDialog: FunctionComponent<RequestDialogProps> = ({
   setIsActive,
   currentHoldNumber,
 }) => {
-  const { offsiteRequesting } = useToggles();
-
-  const availableDates = useAvailableDates();
+  const firstAvailableDate = item.availableDates
+    ? item.availableDates[0]
+    : undefined;
   const [pickUpDate, setPickUpDate] = useState<string | undefined>(
-    availableDates.nextAvailable && dateAsValue(availableDates.nextAvailable)
+    firstAvailableDate && dateAsValue(new Date(firstAvailableDate.from))
   );
 
   function handleConfirmRequest(event: FormEvent<HTMLFormElement>) {
@@ -104,16 +100,7 @@ const RequestDialog: FunctionComponent<RequestDialogProps> = ({
     // these DD-MM-YYYY values, and hopefully they've kept those bugs at bay.
     const pickUpDateValue = pickUpDate ? dateFromValue(pickUpDate) : undefined;
 
-    if (
-      pickUpDateValue &&
-      isRequestableDate({
-        date: pickUpDateValue,
-        startDate: availableDates.nextAvailable,
-        endDate: availableDates.lastAvailable,
-        excludedDates: availableDates.exceptionalClosedDates,
-        excludedDays: availableDates.regularClosedDays,
-      })
-    ) {
+    if (pickUpDateValue) {
       confirmRequest(pickUpDateValue);
     }
   }
@@ -151,36 +138,23 @@ const RequestDialog: FunctionComponent<RequestDialogProps> = ({
             </PickupDeadline>
           </PickUpDateDescription>
           <PickUpDateInputWrapper>
-            <>
-              {offsiteRequesting &&
-                (item.availableDates?.length ? (
-                  <NewRequestingDayPicker
-                    availableDates={item.availableDates}
-                    pickUpDate={pickUpDate}
-                    setPickUpDate={setPickUpDate}
-                  />
-                ) : (
-                  <ErrorMessage>
-                    Error fetching available dates.
-                    <br />
-                    Try again later or email us at
-                    <br />
-                    <a href="mailto:digital@wellcomecollection.org">
-                      digital@wellcomecollection.org
-                    </a>
-                  </ErrorMessage>
-                ))}
-              {!offsiteRequesting && (
-                <RequestingDayPicker
-                  startDate={availableDates.nextAvailable}
-                  endDate={availableDates.lastAvailable}
-                  exceptionalClosedDates={availableDates.exceptionalClosedDates}
-                  regularClosedDays={availableDates.regularClosedDays}
-                  pickUpDate={pickUpDate}
-                  setPickUpDate={setPickUpDate}
-                />
-              )}
-            </>
+            {item.availableDates?.length ? (
+              <RequestingDayPicker
+                availableDates={item.availableDates}
+                pickUpDate={pickUpDate}
+                setPickUpDate={setPickUpDate}
+              />
+            ) : (
+              <ErrorMessage>
+                Error fetching available dates.
+                <br />
+                Try again later or email us at
+                <br />
+                <a href="mailto:digital@wellcomecollection.org">
+                  digital@wellcomecollection.org
+                </a>
+              </ErrorMessage>
+            )}
           </PickUpDateInputWrapper>
         </PickUpDate>
       </Space>
