+++ content/webapp/test/components/ItemRequestModal.test.tsx
@@ -6,14 +6,9 @@ import prismicData from '@weco/common/test/fixtures/prismicData/prismic-data';
 import ItemRequestModal from '../../components/ItemRequestModal/ItemRequestModal';
 import { getItemsWithPhysicalLocation } from '../../utils/works';
 import * as Context from '@weco/common/server-data/Context';
-import * as DateUtils from '../../utils/dates';
 import { renderWithTheme } from '@weco/common/test/fixtures/test-helpers';
 
 jest.spyOn(Context, 'usePrismicData').mockImplementation(() => prismicData);
-
-jest
-  .spyOn(DateUtils, 'determineNextAvailableDate')
-  .mockImplementation(() => new Date('2022-05-21'));
 
 const RequestModal: FunctionComponent = () => {
   const [requestModalIsActive, setRequestModalIsActive] = useState(true);
@@ -36,7 +31,7 @@ const RequestModal: FunctionComponent = () => {
 };
 
 describe('ItemRequestModal', () => {
-  it('allows an available date to selected (based on business rules about what dates for collection should be available) and displays the entered date', async () => {
+  it('allows an available date (as returned with the item from the itemsAPI) to be selected, and displays the entered date', async () => {
     const { getByLabelText } = renderWithTheme(<RequestModal />);
     const select = getByLabelText(/^Select a date$/i) as HTMLSelectElement;
     await act(async () => {
